Extracted out a generic string utils class
